requirements.txt
absl-py==0.14.1
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
decorator==5.1.0
Distance==0.1.3
docopt==0.6.2
filelock==3.3.0
Flask==2.0.2
flatbuffers==1.12
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.4.0
german-transliterate @ git+https://github.com/repodiac/german_transliterate@4e01beba5f19adb9ccd72d0607a81f4ed7f6ee87
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==2.10.0
heroku==0.1.4
huggingface-hub==0.0.19
idna==3.3
importlib-metadata==4.8.1
inflect==5.3.0
itsdangerous==2.0.1
jamo==0.4.1
Jinja2==3.0.2
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
nltk==3.6.5
num2words==0.5.10
numba==0.54.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.2
pooch==1.5.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pypinyin==0.43.0
python-dateutil==1.5
pyworld==0.3.0
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
sic==1.3.3
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
TextGrid==1.5
threadpoolctl==3.0.0
tqdm==4.62.3
typeguard==2.13.0
typing-extensions==3.7.4.3
unicode==2.8
Unidecode==1.3.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
